requirements.txt
Flask==3.0.0
Pillow==10.1.0
reportlab==4.0.7
gunicorn==21.2.0